Reorganize Readme & remove unnecessary docs
